all tests passing with new structure for erc20 token storage
